benchmark/lightweight-openpose/requirements.txt
addict==2.4.0
matplotlib==3.2.2
mmcv==1.4.3
mmcv_full==1.4.1
mmpose==0.21.0
nni==2.5
numpy==1.21.2
opencv_python==4.5.4.60
pycocotools==2.0.0
PyYAML==6.0
scikit_image==0.19.1
skimage==0.0
thop==0.0.31.post2005241907
torch==1.10.1
torchvision==0.11.2
tqdm==4.61.2
